# api/openapi.yml
openapi: '3.0.2'

info:
  title: "ToDoApp"
  version: '1.0'
servers:
  - url: "https://localhost:8080"
    description: "local"
    
paths:
  /tasks:
    get:
      summary: get all tasks
      description: Get title, is_completed and deadline from tasks table in json format
      operationId: GetTaskList
      tags:
        - Todo List
      deprecated: false
      responses:
        200:
          description: Got data

  /tasks/{id}:
    get:
      summary: get specific task
      description: Get columns of a specific task from tasks table in json format. the columns are id, title, content, is_completed and deadline
      operationId: GetSpecificTask
      tags:
        - Task Detail
      parameters:
        - name: id
          in: path
          required: true
          schema: 
            type: integer
          description: id in tasks table
      deprecated: false
      responses:
        200:
          description: Got data

  /create_task:
    post:
      summary: create tasks
      description: create tasks data from inputted data
      operationId: CreateTask
      tags:
        - Create Tasks
      deprecated: false
      responses:
        200:
          description: Created data

  /update-task/{id}:
    put:
      summary: update specific task
      description: Update existing task in tasks table. Get task by id.
      operationId: UpdateTask
      tags:
        - Update Task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: id in tasks table
      deprecated: false
      responses:
        200:
          description: Updated data

  /delete-task/{id}:
    delete:
      summary: delete specific task
      description: Delete existing task in tasks table. Get task by id.
      operationId: DeleteTask
      tags:
        - Delete Task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: id in tasks table
      deprecated: false
      responses:
        200:
          description: :Deleted data